+++ src/main/java/org/realityforge/bazel/depgen/config/JspecifyMode.java
@@ -1,0 +1,7 @@
+package org.realityforge.bazel.depgen.config;
+
+public enum JspecifyMode {
+    Autodetect,
+    Enable,
+    Disable
+}